chore: move profile form into separate component
