+++ src/css/base.scss
@@ -4130,3 +4130,14 @@ textarea::placeholder {
     box-shadow: 0px 4px 8px 0px rgba(0, 0, 0, 0.35);
     backdrop-filter: blur(5px);
 }
+
+
+// Tab behaviors
+.dc__tab-focus {
+    outline: none;
+}
+
+.dc__tab-focus:focus-visible {
+    outline: 5px auto -webkit-focus-ring-color;
+    outline-offset: -2px;
+}
